@turol 2024-08-10
I managed to fix UBSAN. It turns out that sometimes `ND_SIGN_EX` is called with `sz == 0` ...