@turol 2024-08-22
I/O effects should have been eliminated by the pre-runs before the actual benchmark loop. Interrupt effects can't be eliminated without radically changing the system configuration. If the OS wants to punt our program from the CPU to do something else it will. We can only minimize the effect by running more samples. 

I replaced use of `time` with instructions/second from `-iv` output. Now higher value is better. However I still get the same result.

You appear to be using `clock()` for measuring time without using `CLOCKS_PER_SEC`
https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/clock?view=msvc-170
So some of my numbers are different from yours by a factor of 1000.

Numbers on my Intel i7-12700H laptop but I get broadly similar results from AMD 5950x.

For a real executable (.text section only, 32MB):
```
Disassembled 19791808 instructions in 1156758ms, 17109.7222 instructions/second, 141.391124 clocks/instruction, average ilen 3.393740 bytes
Disassembled 19791808 instructions in 1094370ms, 18085.1156 instructions/second, 132.392333 clocks/instruction, average ilen 3.393740 bytes

    N           Min           Max        Median           Avg        Stddev
x  30     16955.914     17147.389     17066.273     17059.275     50.281244
+  30     17805.978     18094.127      18026.59     18007.061     63.651212
Difference at 95.0% confidence
	947.786 +/- 29.6487
	5.55584% +/- 0.173798%
	(Student's t, pooled s = 57.3571)
```

And 32MB of randomness:
```
Disassembled 12987011 instructions in 856135ms, 15169.3495 instructions/second, 161.957771 clocks/instruction, average ilen 2.822512 bytes
Disassembled 12987011 instructions in 809213ms, 16048.9401 instructions/second, 152.054722 clocks/instruction, average ilen 2.822512 bytes

    N           Min           Max        Median           Avg        Stddev
x  30     14934.328      15169.35     15072.905      15062.49     58.431712
+  30      15013.99      16048.94     15972.651      15932.42     181.52074
Difference at 95.0% confidence
	869.93 +/- 69.7011
	5.77547% +/- 0.462746%
	(Student's t, pooled s = 134.841)
```
There is something very very strange going on in your setup.
Are you also using WSL2? Can you try on real non-virtualized setup?

Are you only running each benchmark once?  The reason I use a multiple runs is to
1. Get some proper statistics for ministat
2. Alternate between the two programs to reduce the effect of clock speed rampup or thermal throttling
3. Reduce the effect of virtual memory layout