@vlutas 2024-08-22
There is nothing strange about my setup - it's bare metal Windows. 
What is different though is that you're running the tests on Linux (WSL), while I'm running them on Windows. Ergo, it probably boils down to how the compiler decides to generate & arrange the code (GCC vs. MSVC), as you're reporting improvement, whereas I have a downgrade of performance. We may both be right... :) 
However, like I said in my previous comment, I currently fail to see a **consistent** performance improvement with these changes. I will investigate how it behaves in WSL as well, and I will go through the code changes, to see what should be kept, and what should be modified in order to get a consistent performance improvement.
Thank you again for your contribution!